chore(python-client): rename python package to `streampipes`
