+++ streampipes-client-python/docs/index.md
@@ -36,18 +36,9 @@ and the amazing universe of data analytics libraries in Python. </p>
 
 <br>
 
-** ❗❗❗IMPORTANT ❗❗❗**
-<br>
-<br>
-**The current version of this Python client is still in alpha phase at best.**
+**💡 The current version of this Python client is still a beta version.**
 <br>
 **This means that it is still heavily under development, which may result in frequent and extensive API changes, unstable behavior, etc.**
-<br>
-**Please consider it only as a sneak preview.**
-<br>
-<br>
-**❗❗❗ IMPORTANT ❗❗❗**
-
 <br>
 
 ## ⚡️ Quickstart
@@ -56,33 +47,45 @@ As a quick example, we demonstrate how to set up and configure a StreamPipes cli
 In addition, we will get the available data lake measures out of StreamPipes.
 
 ```python
->>> from streampipes_client.client import StreamPipesClient
->>> from streampipes_client.client.client_config import StreamPipesClientConfig
->>> from streampipes_client.client.credential_provider import StreamPipesApiKeyCredentials
+>> > from streampipes.client import StreamPipesClient
+>> > from streampipes.client.client_config import StreamPipesClientConfig
+>> > from streampipes.client.credential_provider import StreamPipesApiKeyCredentials
 
->>> config = StreamPipesClientConfig(
-...    credential_provider=StreamPipesApiKeyCredentials(
-...         username="test@streampipes.apache.org",
-...         api_key="DEMO-KEY",
-...         ),
-...     host_address="localhost",
-...     http_disabled=True,
-...     port=80                  
+>> > config = StreamPipesClientConfig(
+    ...
+credential_provider = StreamPipesApiKeyCredentials(
+    ...
+username = "test@streampipes.apache.org",
+           ...
+api_key = "DEMO-KEY",
+          ...         ),
+...
+host_address = "localhost",
+               ...
+http_disabled = True,
+                ...
+port = 80
 ...)
 
->>> client = StreamPipesClient(client_config=config)
+>> > client = StreamPipesClient(client_config=config)
 
 # get all available datat lake measures
->>> measures = client.dataLakeMeasureApi.all()
+>> > measures = client.dataLakeMeasureApi.all()
 
 # get amount of retrieved measures
->>> len(measures)
+>> > len(measures)
 1
 
 # inspect the data lake measures as pandas dataframe
->>> measures.to_pandas()
-    measure_name timestamp_field  ... pipeline_is_running num_event_properties
-0           test   s0::timestamp  ...               False                    2
+>> > measures.to_pandas()
+measure_name
+timestamp_field...pipeline_is_running
+num_event_properties
+0
+test
+s0::timestamp...
+False
+2
 [1 rows x 6 columns]
 ```
 <br>
@@ -90,7 +93,7 @@ Alternatively, you can provide your credentials via environment variables.
 Simply define your credential provider as follows:
 
 ```python
->>> from streampipes_client.client.credential_provider import StreamPipesApiKeyCredentials
+>> > from streampipes.client.credential_provider import StreamPipesApiKeyCredentials
 
 StreamPipesApiKeyCredentials.from_env(username_env="USER", api_key_env="API-KEY")
 ```
