+++ streampipes-client-python/streampipes/__version__.py
