+++ streampipes-client-python/streampipes/client/credential_provider.py
