+++ streampipes-client-python/streampipes/endpoint/api/data_stream.py
@@ -20,14 +20,14 @@ Specific implementation of the StreamPipes API's data stream endpoints.
 """
 from typing import Tuple, Type
 
-from streampipes_client.endpoint.endpoint import APIEndpoint
-from streampipes_client.model.container import DataStreams
+from streampipes.endpoint.endpoint import APIEndpoint
+from streampipes.model.container import DataStreams
 
 __all__ = [
     "DataStreamEndpoint",
 ]
 
-from streampipes_client.model.container.resource_container import ResourceContainer
+from streampipes.model.container.resource_container import ResourceContainer
 
 
 class DataStreamEndpoint(APIEndpoint):
@@ -44,9 +44,9 @@ class DataStreamEndpoint(APIEndpoint):
     Examples
     --------
 
-    >>> from streampipes_client.client import StreamPipesClient
-    >>> from streampipes_client.client.client_config import StreamPipesClientConfig
-    >>> from streampipes_client.client.credential_provider import StreamPipesApiKeyCredentials
+    >>> from streampipes.client import StreamPipesClient
+    >>> from streampipes.client.client_config import StreamPipesClientConfig
+    >>> from streampipes.client.credential_provider import StreamPipesApiKeyCredentials
 
     >>> client_config = StreamPipesClientConfig(
     ...     credential_provider=StreamPipesApiKeyCredentials(username="test-user", api_key="api-key"),
