+++ streampipes-client-python/streampipes/endpoint/exceptions.py
