+++ streampipes-client-python/streampipes/functions/utils/async_iter_handler.py
