+++ streampipes-client-python/streampipes/model/common.py
