+++ streampipes-client-python/streampipes/model/resource/__init__.py
