+++ streampipes-client-python/tests/functions/test_function_handler.py
@@ -19,13 +19,13 @@ from typing import Any, Dict, List, Tuple
 from unittest import TestCase
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from streampipes_client.client.client import StreamPipesClient, StreamPipesClientConfig
-from streampipes_client.client.credential_provider import StreamPipesApiKeyCredentials
-from streampipes_client.functions.function_handler import FunctionHandler
-from streampipes_client.functions.registration import Registration
-from streampipes_client.functions.streampipes_function import StreamPipesFunction
-from streampipes_client.functions.utils.function_context import FunctionContext
-from streampipes_client.model.resource.data_stream import DataStream
+from streampipes.client.client import StreamPipesClient, StreamPipesClientConfig
+from streampipes.client.credential_provider import StreamPipesApiKeyCredentials
+from streampipes.functions.function_handler import FunctionHandler
+from streampipes.functions.registration import Registration
+from streampipes.functions.streampipes_function import StreamPipesFunction
+from streampipes.functions.utils.function_context import FunctionContext
+from streampipes.model.resource.data_stream import DataStream
 
 
 class TestFunction(StreamPipesFunction):
@@ -206,11 +206,11 @@ class TestFunctionHandler(TestCase):
             {"density": 3.6, "temperature": 30.4, "timestamp": 1670000007000},
         ]
 
-    @patch("streampipes_client.functions.function_handler.NatsBroker.disconnect", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker._createSubscription", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker._makeConnection", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker.get_message", autospec=True)
-    @patch("streampipes_client.client.client.Session", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker.disconnect", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker._createSubscription", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker._makeConnection", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker.get_message", autospec=True)
+    @patch("streampipes.client.client.Session", autospec=True)
     def test_function_handler(self, http_session: MagicMock, nats_broker: MagicMock, *args: Tuple[AsyncMock]):
         http_session_mock = MagicMock()
         http_session_mock.get.return_value.json.return_value = self.data_stream
@@ -241,11 +241,11 @@ class TestFunctionHandler(TestCase):
         self.assertListEqual(test_function.data, self.test_stream_data1)
         self.assertTrue(test_function.stopped)
 
-    @patch("streampipes_client.functions.function_handler.NatsBroker.disconnect", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker._createSubscription", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker._makeConnection", autospec=True)
-    @patch("streampipes_client.functions.function_handler.NatsBroker.get_message", autospec=True)
-    @patch("streampipes_client.endpoint.endpoint.APIEndpoint.get", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker.disconnect", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker._createSubscription", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker._makeConnection", autospec=True)
+    @patch("streampipes.functions.function_handler.NatsBroker.get_message", autospec=True)
+    @patch("streampipes.endpoint.endpoint.APIEndpoint.get", autospec=True)
     def test_function_handler_two_streams(self, endpoint: MagicMock, nats_broker: MagicMock, *args: Tuple[AsyncMock]):
         def get_stream(endpoint, stream_id):
             if stream_id == "urn:streampipes.apache.org:eventstream:uPDKLI":
